♻️ Refactor: OnBoarding 관련 MVI 클래스 패키지 위치 변경 (`model` -> `contract`)
