+++ feature/onboarding/src/main/java/com/twix/onboarding/dday/DdayRoute.kt
@@ -32,10 +32,10 @@ import com.twix.designsystem.theme.GrayColor
 import com.twix.designsystem.theme.TwixTheme
 import com.twix.domain.model.enums.AppTextStyle
 import com.twix.onboarding.OnBoardingViewModel
+import com.twix.onboarding.contract.OnBoardingIntent
+import com.twix.onboarding.contract.OnBoardingSideEffect
 import com.twix.onboarding.dday.component.DDayField
 import com.twix.onboarding.dday.component.DdayTopBar
-import com.twix.onboarding.model.OnBoardingIntent
-import com.twix.onboarding.model.OnBoardingSideEffect
 import com.twix.ui.base.ObserveAsEvents
 import org.koin.compose.koinInject
 import java.time.LocalDate
